Create Middlewares: CheckTagOwnership, CheckSubtaskOwnership and CheckTagOwnership

+++ routes/api.php
@@ -4,6 +4,9 @@ use App\Http\Controllers\API\AuthController;
 use App\Http\Controllers\API\Subtask\SubtaskController;
 use App\Http\Controllers\API\Tags\TagController;
 use App\Http\Controllers\API\Tasks\TaskController;
+use App\Http\Middleware\CheckSubtaskOwnership;
+use App\Http\Middleware\CheckTagOwnership;
+use App\Http\Middleware\CheckTaskOwnership;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -18,26 +21,26 @@ Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanc
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('/tasks', [TaskController::class, 'index']);
     Route::post('/tasks', [TaskController::class, 'store']);
-    Route::get('/tasks/{task_id}', [TaskController::class, 'show']);
-    Route::patch('/tasks/{task_id}', [TaskController::class, 'update']);
-    Route::delete('/tasks/{task_id}', [TaskController::class, 'destroy']);
+    Route::get('/tasks/{task_id}', [TaskController::class, 'show'])->middleware(CheckTaskOwnership::class);
+    Route::patch('/tasks/{task_id}', [TaskController::class, 'update'])->middleware(CheckTaskOwnership::class);
+    Route::delete('/tasks/{task_id}', [TaskController::class, 'destroy'])->middleware(CheckTaskOwnership::class);
     Route::get('/tasks/priority/{priority}', [TaskController::class, 'tasksByPriority']);
 });
 
 Route::middleware('auth:sanctum')->group(function () {
-    Route::get('/tasks/{task_id}/subtasks', [SubtaskController::class, 'index']);
-    Route::post('/tasks/{task_id}/subtasks', [SubtaskController::class, 'store']);
-    Route::get('/tasks/{task_id}/subtasks/{subtask_id}', [SubtaskController::class, 'show']);
-    Route::patch('/tasks/{task_id}/subtasks/{subtask_id}', [SubtaskController::class, 'update']);
-    Route::delete('/tasks/{task_id}/subtasks/{subtask_id}', [SubtaskController::class, 'destroy']);
+    Route::get('/tasks/{task_id}/subtasks', [SubtaskController::class, 'index'])->middleware(CheckTaskOwnership::class);
+    Route::post('/tasks/{task_id}/subtasks', [SubtaskController::class, 'store'])->middleware(CheckTaskOwnership::class);
+    Route::get('/tasks/{task_id}/subtasks/{subtask_id}', [SubtaskController::class, 'show'])->middleware(CheckSubtaskOwnership::class);
+    Route::patch('/tasks/{task_id}/subtasks/{subtask_id}', [SubtaskController::class, 'update'])->middleware(CheckSubtaskOwnership::class);
+    Route::delete('/tasks/{task_id}/subtasks/{subtask_id}', [SubtaskController::class, 'destroy'])->middleware(CheckSubtaskOwnership::class);
 });
 
 
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('/tags', [TagController::class, 'index']);
     Route::post('/tags', [TagController::class, 'store']);
-    Route::get('/tags/{tag_id}', [TagController::class, 'show']);
-    Route::patch('/tags/{tag_id}', [TagController::class, 'update']);
-    Route::delete('/tags/{tag_id}', [TagController::class, 'destroy']);
-    Route::get('/tags/{tag_id}/tasks', [TagController::class, 'tasks']);
+    Route::get('/tags/{tag_id}', [TagController::class, 'show'])->middleware(CheckTagOwnership::class);
+    Route::patch('/tags/{tag_id}', [TagController::class, 'update'])->middleware(CheckTagOwnership::class);
+    Route::delete('/tags/{tag_id}', [TagController::class, 'destroy'])->middleware(CheckTagOwnership::class);
+    Route::get('/tags/{tag_id}/tasks', [TagController::class, 'tasks'])->middleware(CheckTagOwnership::class);
 });
